+++ R/zzz.R
@@ -4,7 +4,8 @@
 # see https://github.com/hadley/r-pkgs/blob/master/r.rmd
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("\nNOTE: install ggplot2 to run coor.graph()")
+ if(!requireNamespace("ggplot2",quietly = TRUE))
+   packageStartupMessage("\nNOTE: install ggplot2 to run coor.graph()")
 }
 
 .onLoad <- function(libname, pkgname) {